requirements.txt
lxml==6.0.0
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
pytesseract==0.3.13
python-docx==1.2.0
typing_extensions==4.14.1
